git-svn-id: https://active-ist.svn.sourceforge.net/svnroot/active-ist/libraries/jaus++/trunk@147 846840b6-6447-0410-a3b3-dec9255eb4eb

+++ 1.0/src/jaus/video/visualsensor.cpp
@@ -355,7 +355,7 @@ int VisualSensor::SetCurrentFrameCompressed(const Byte* image,
         memcpy(mpCompressedImage, image, size);
         mCompressedSize = size;
         mFrameNumber++;
-        mCompressionNumber = mFrameNumber;
+        mCompressionNumber = (UShort)mFrameNumber;
         if(decompress || mEnableSharedImageFlag)
         {
             if(mRawImage.Decompress(image, size, format))
